Share Rust sandbox image builder across SDKs
