+++ crates/cloud-sdk/Cargo.toml
@@ -11,7 +11,10 @@ documentation = "https://docs.rs/tensorlake"
 
 [dependencies]
 bytes = { workspace = true }
+base64 = { workspace = true }
+bollard = { workspace = true }
 chrono = { workspace = true }
+docker-credentials-config = { workspace = true }
 utoipa = { workspace = true, optional = true }
 derive_builder = { workspace = true }
 flate2 = { workspace = true }
@@ -29,6 +32,7 @@ serde_path_to_error = { workspace = true }
 serde_repr = { workspace = true }
 serde_with = { workspace = true }
 sha2 = { workspace = true }
+shlex = { workspace = true }
 tar = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
