+++ crates/rust-cloud-sdk-py/src/lib.rs
@@ -7,6 +7,7 @@
 use std::error::Error as StdError;
 use std::future::Future;
 use std::io::Write;
+use std::path::PathBuf;
 use std::time::Duration;
 
 use futures::StreamExt;
@@ -2605,6 +2606,77 @@ fn create_image_context_file(
     Ok(())
 }
 
+#[pyfunction]
+#[pyo3(signature = (
+    api_url,
+    token,
+    dockerfile_path,
+    registered_name=None,
+    disk_mb=None,
+    builder_disk_mb=None,
+    cpus=None,
+    memory_mb=None,
+    is_public=false,
+    organization_id=None,
+    project_id=None,
+    namespace=None,
+    use_scope_headers=false,
+    user_agent=None,
+))]
+fn build_sandbox_image(
+    api_url: String,
+    token: String,
+    dockerfile_path: String,
+    registered_name: Option<String>,
+    disk_mb: Option<u64>,
+    builder_disk_mb: Option<u64>,
+    cpus: Option<f64>,
+    memory_mb: Option<i64>,
+    is_public: bool,
+    organization_id: Option<String>,
+    project_id: Option<String>,
+    namespace: Option<String>,
+    use_scope_headers: bool,
+    user_agent: Option<String>,
+) -> PyResult<String> {
+    let options = tensorlake::sandbox_images::SandboxImageBuildOptions {
+        api_url,
+        bearer_token: token,
+        use_scope_headers,
+        organization_id,
+        project_id,
+        namespace: namespace.unwrap_or_else(|| "default".to_string()),
+        dockerfile_path: PathBuf::from(dockerfile_path),
+        registered_name,
+        disk_mb,
+        builder_disk_mb,
+        cpus,
+        memory_mb,
+        is_public,
+        user_agent,
+    };
+
+    let result = shared_runtime()
+        .block_on(
+            async move { tensorlake::sandbox_images::build_sandbox_image(options, |_| {}).await },
+        )
+        .map_err(|error| {
+            CloudSandboxClientError::new_err((
+                "sandbox_image_build",
+                Option::<u16>::None,
+                error.to_string(),
+            ))
+        })?;
+
+    serde_json::to_string(&result).map_err(|error| {
+        CloudSandboxClientError::new_err((
+            "sandbox_image_build",
+            Option::<u16>::None,
+            error.to_string(),
+        ))
+    })
+}
+
 #[pymodule]
 fn _cloud_sdk(_py: Python<'_>, module: &Bound<'_, PyModule>) -> PyResult<()> {
     // Eager-init the shared tokio runtime at import time so the first sync
@@ -2627,5 +2699,6 @@ fn _cloud_sdk(_py: Python<'_>, module: &Bound<'_, PyModule>) -> PyResult<()> {
     module.add_class::<CloudSandboxDesktopClient>()?;
     module.add_class::<CloudDocumentAIClient>()?;
     module.add_function(wrap_pyfunction!(create_image_context_file, module)?)?;
+    module.add_function(wrap_pyfunction!(build_sandbox_image, module)?)?;
     Ok(())
 }
