+++ src/tensorlake/image/image.py
@@ -104,6 +104,7 @@ class Image:
         cpus: float = 2.0,
         memory_mb: int = 4096,
         disk_mb: int | None = None,
+        builder_disk_mb: int | None = None,
         is_public: bool = False,
         context_dir: str | None = None,
         verbose: bool = False,
@@ -117,7 +118,8 @@ class Image:
             registered_name: Name to register the image under. Defaults to ``self.name``.
             cpus: CPUs for the build sandbox (default 2.0).
             memory_mb: Memory for the build sandbox in MB (default 4096).
-            disk_mb: Root disk size for the build sandbox in MB.
+            disk_mb: Root disk size for the generated sandbox image in MB.
+            builder_disk_mb: Root disk size for the temporary builder sandbox in MB.
             is_public: Make the registered image publicly accessible.
             context_dir: Directory used to resolve relative COPY/ADD paths.
                 Defaults to the current working directory.
@@ -134,6 +136,7 @@ class Image:
             cpus=cpus,
             memory_mb=memory_mb,
             disk_mb=disk_mb,
+            builder_disk_mb=builder_disk_mb,
             is_public=is_public,
             context_dir=context_dir,
             verbose=verbose,
